+++ src/spyglass/decoding/v1/sorted_spikes.py
@@ -275,8 +275,8 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
         """Retrieve the decoding model"""
         return SortedSpikesDetector.load_model(self.fetch1("classifier_path"))
 
-    @staticmethod
-    def fetch_environments(key):
+    @classmethod
+    def fetch_environments(cls, key):
         """Fetch the environments for the decoding model
 
         Parameters
@@ -289,6 +289,10 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
         List[TrackGraph]
             list of track graphs in the trained model
         """
+        key = cls.get_fully_defined_key(
+            key, required_fields=["decoding_param_name"]
+        )
+
         model_params = (
             DecodingParameters
             & {"decoding_param_name": key["decoding_param_name"]}
@@ -314,8 +318,8 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
 
         return classifier.environments
 
-    @staticmethod
-    def fetch_position_info(key):
+    @classmethod
+    def fetch_position_info(cls, key):
         """Fetch the position information for the decoding model
 
         Parameters
@@ -328,6 +332,16 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
         Tuple[pd.DataFrame, List[str]]
             The position information and the names of the position variables
         """
+        key = cls.get_fully_defined_key(
+            key,
+            required_fields=[
+                "position_group_name",
+                "nwb_file_name",
+                "encoding_interval",
+                "decoding_interval",
+            ],
+        )
+
         position_group_key = {
             "position_group_name": key["position_group_name"],
             "nwb_file_name": key["nwb_file_name"],
@@ -339,8 +353,8 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
 
         return position_info, position_variable_names
 
-    @staticmethod
-    def fetch_linear_position_info(key):
+    @classmethod
+    def fetch_linear_position_info(cls, key):
         """Fetch the position information and project it onto the track graph
 
         Parameters
@@ -353,6 +367,16 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
         pd.DataFrame
             The linearized position information
         """
+        key = cls.get_fully_defined_key(
+            key,
+            required_fields=[
+                "position_group_name",
+                "nwb_file_name",
+                "encoding_interval",
+                "decoding_interval",
+            ],
+        )
+
         environment = SortedSpikesDecodingV1.fetch_environments(key)[0]
 
         position_df = SortedSpikesDecodingV1.fetch_position_info(key)[0]
@@ -374,9 +398,13 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
             axis=1,
         ).loc[min_time:max_time]
 
-    @staticmethod
+    @classmethod
     def fetch_spike_data(
-        key, filter_by_interval=True, time_slice=None, return_unit_ids=False
+        cls,
+        key,
+        filter_by_interval=True,
+        time_slice=None,
+        return_unit_ids=False,
     ) -> Union[list[np.ndarray], Optional[list[dict]]]:
         """Fetch the spike times for the decoding model
 
@@ -399,6 +427,14 @@ class SortedSpikesDecodingV1(SpyglassMixin, dj.Computed):
         list[np.ndarray]
             List of spike times for each unit in the model's spike group
         """
+        key = cls.get_fully_defined_key(
+            key,
+            required_fields=[
+                "encoding_interval",
+                "decoding_interval",
+            ],
+        )
+
         spike_times, unit_ids = SortedSpikesGroup.fetch_spike_data(
             key, return_unit_ids=True
         )