+++ src/spyglass/spikesorting/analysis/v1/group.py
@@ -3,7 +3,6 @@ from typing import Optional, Union
 
 import datajoint as dj
 import numpy as np
-from ripple_detection import get_multiunit_population_firing_rate
 
 from spyglass.common import Session  # noqa: F401
 from spyglass.settings import test_mode
@@ -127,9 +126,12 @@ class SortedSpikesGroup(SpyglassMixin, dj.Manual):
             include_mask[ind] = True
         return include_mask
 
-    @staticmethod
+    @classmethod
     def fetch_spike_data(
-        key: dict, time_slice: list[float] = None, return_unit_ids: bool = False
+        cls,
+        key: dict,
+        time_slice: Union[list[float], slice] = None,
+        return_unit_ids: bool = False,
     ) -> Union[list[np.ndarray], Optional[list[dict]]]:
         """fetch spike times for units in the group
 
@@ -137,7 +139,7 @@ class SortedSpikesGroup(SpyglassMixin, dj.Manual):
         ----------
         key : dict
             dictionary containing the group key
-        time_slice : list of float, optional
+        time_slice : list of float or slice, optional
             if provided, filter for spikes occurring in the interval [start, stop], by default None
         return_unit_ids : bool, optional
             if True, return the unit_ids along with the spike times, by default False
@@ -148,6 +150,8 @@ class SortedSpikesGroup(SpyglassMixin, dj.Manual):
         list of np.ndarray
             list of spike times for each unit in the group
         """
+        key = cls.get_fully_defined_key(key)
+
         # get merge_ids for SpikeSortingOutput
         merge_ids = (
             (
@@ -185,8 +189,14 @@ class SortedSpikesGroup(SpyglassMixin, dj.Manual):
             ]
 
             # filter the spike times based on the labels if present
-            if "label" in nwb_file[nwb_field_name]:
-                group_label_list = nwb_file[nwb_field_name]["label"].to_list()
+            for col in ("label", "curation_label"):  # v0, v1 names
+                if col in nwb_file[nwb_field_name].columns:
+                    group_labels = nwb_file[nwb_field_name][col]
+                    break
+            else:
+                group_labels = None
+            if group_labels is not None:
+                group_label_list = group_labels.to_list()
                 include_unit = SortedSpikesGroup.filter_units(
                     group_label_list, include_labels, exclude_labels
                 )
@@ -198,6 +208,8 @@ class SortedSpikesGroup(SpyglassMixin, dj.Manual):
 
             # filter the spike times based on the time slice if provided
             if time_slice is not None:
+                if isinstance(time_slice, (list, tuple)):
+                    time_slice = slice(*time_slice)
                 sorting_spike_times = [
                     times[
                         np.logical_and(